==> analysis/01_Humans_GPT4_and_CA_performances/03_create_GPT4_matrices.R <==
# This code create the three dimensional array corresponding to GPT4+LCA, GPT4+SCA and GPT4+DCA classifications from the multiple txt files generated by each agent.

# Store the possible agents
agent = c("1","2","3")

# Create the standard array skeleton that will be the storage template
matrix_result_template = array(data=0, dim=c(
  length(index),
  length(category),
  length(tone),
  length(agent)
), 
dimnames = list(index,category,tone,agent))


# -------------------------------------------------------------------------------------------------------------
# step 1 : GPT-4 alone
# -------------------------------------------------------------------------------------------------------------

# Create the array skeleton for GPT4 standalone
GPT4 = matrix_result_template

# Txt files generated by the prediction phase are always valid json files.
for(current_agent in 1:length(agent)){

  # Listing the paths
  file_paths = data.frame(
    index = index,
    path = paste0("analysis/01_Humans_GPT4_and_CA_performances/data/02_GPT4/output_gpt4-",current_agent,"/llm_queries/output_1/initial_classification_result_",index,".json")
  )
  
  for(i in 1:length(index)){
    
    # Gather the LLM classification :
    suppressWarnings({
      result = paste(trimws(readLines(file_paths$path[i])),collapse=" ")
    })
    result = jsonlite::fromJSON(result)
    
    # Extract the output from the result
    result = result$output
    
    # erase the themes, keep the categories
    result_2 = list() # json data, but without themes
    themes = ls(result)
    for(current_theme in themes){
      result_2 = append(result_2, result[[current_theme]])
    }
    names(result_2) = trimws(names(result_2)) # Avoid white spaces added by LLM
    result = result_2 # Save the trimmed json in result
    
    # Store the prediction in the 3D matrix
    for(current_category in category){
      GPT4[i,current_category,"positive",current_agent] = ifelse(result[[current_category]][["positive"]]!="",1,0)
      GPT4[i,current_category,"negative",current_agent] = ifelse(result[[current_category]][["negative"]]!="",1,0)
    }
  }
}

saveRDS(GPT4,"analysis/01_Humans_GPT4_and_CA_performances/data/02_GPT4/GPT4.rds")



# -------------------------------------------------------------------------------------------------------------
# step 2 : GPT-4 + SCA
# -------------------------------------------------------------------------------------------------------------

# Create the array skeleton for GPT4_SCA
GPT4_SCA = matrix_result_template

# Txt files generated by the prediction phase are always valid json files.
for(current_agent in 1:length(agent)){
  
  # Listing the paths
  file_paths = data.frame(
    index = index,
    path = paste0("analysis/01_Humans_GPT4_and_CA_performances/data/02_GPT4/output_gpt4-",current_agent,"/evaluations/sca/result_",index,".json")
  )
  
  for(i in 1:length(index)){
    
    # Gather the LLM classification :
    suppressWarnings({
      result = paste(trimws(readLines(file_paths$path[i])),collapse=" ")
    })
    result = jsonlite::fromJSON(result)
    
    # Extract the output from the result
    result = result$output
    
    # erase the themes, keep the categories
    result_2 = list() # json data, but without themes
    themes = ls(result)
    for(current_theme in themes){
      result_2 = append(result_2, result[[current_theme]])
    }
    names(result_2) = trimws(names(result_2)) # Avoid white spaces added by LLM
    result = result_2 # Save the trimmed json in result
    
    # Store the prediction in the 3D matrix
    for(current_category in category){
      GPT4_SCA[i,current_category,"positive",current_agent] = result[[current_category]][["positive"]]
      GPT4_SCA[i,current_category,"negative",current_agent] = result[[current_category]][["negative"]]
    }
  }
}

saveRDS(GPT4_SCA,"analysis/01_Humans_GPT4_and_CA_performances/data/02_GPT4/GPT4_SCA.rds")



# -------------------------------------------------------------------------------------------------------------
# step 3 : GPT-4 + LCA
# -------------------------------------------------------------------------------------------------------------

# Create the array skeleton for GPT4+LCA
GPT4_LCA = matrix_result_template

# Txt files generated by the prediction phase are always valid json files.
for(current_agent in 1:length(agent)){
  
  # Listing the paths
  file_paths = data.frame(
    index = index,
    path = paste0("analysis/01_Humans_GPT4_and_CA_performances/data/02_GPT4/output_gpt4-",current_agent,"/evaluations/lca/output_1/result_",index,".json")
  )
  
  for(i in 1:length(index)){
    
    # Gather the LLM classification :
    suppressWarnings({
      result = paste(trimws(readLines(file_paths$path[i])),collapse=" ")
    })
    result = jsonlite::fromJSON(result)
    
    # Extract the output from the result
    result = result$output
    
    # erase the themes, keep the categories
    result_2 = list() # json data, but without themes
    themes = ls(result)
    for(current_theme in themes){
      result_2 = append(result_2, result[[current_theme]])
    }
    names(result_2) = trimws(names(result_2)) # Avoid white spaces added by LLM
    result = result_2 # Save the trimmed json in result
    
    # Store the prediction in the 3D matrix
    for(current_category in category){
      GPT4_LCA[i,current_category,"positive",current_agent] = result[[current_category]][["positive"]]
      GPT4_LCA[i,current_category,"negative",current_agent] = result[[current_category]][["negative"]]
    }
  }
}

saveRDS(GPT4_LCA,"analysis/01_Humans_GPT4_and_CA_performances/data/02_GPT4/GPT4_LCA.rds")



# -------------------------------------------------------------------------------------------------------------
# step 4 : GPT-4 + SLCA
# -------------------------------------------------------------------------------------------------------------

# Create the array skeleton for GPT4+SLCA
GPT4_SLCA = matrix_result_template

# Txt files generated by the prediction phase are always valid json files.
for(current_agent in 1:length(agent)){
  
  # Listing the paths
  file_paths = data.frame(
    index = index,
    path = paste0("analysis/01_Humans_GPT4_and_CA_performances/data/02_GPT4/output_gpt4-",current_agent,"/evaluations/slca/result_",index,".json")
  )
  
  for(i in 1:length(index)){
    
    # Gather the LLM classification :
    suppressWarnings({
      result = paste(trimws(readLines(file_paths$path[i])),collapse=" ")
    })
    result = jsonlite::fromJSON(result)
    
    # Extract the output from the result
    result = result$output
    
    # erase the themes, keep the categories
    result_2 = list() # json data, but without themes
    themes = ls(result)
    for(current_theme in themes){
      result_2 = append(result_2, result[[current_theme]])
    }
    names(result_2) = trimws(names(result_2)) # Avoid white spaces added by LLM
    result = result_2 # Save the trimmed json in result
    
    # Store the prediction in the 3D matrix
    for(current_category in category){
      GPT4_SLCA[i,current_category,"positive",current_agent] = result[[current_category]][["positive"]]
      GPT4_SLCA[i,current_category,"negative",current_agent] = result[[current_category]][["negative"]]
    }
  }
}

saveRDS(GPT4_SLCA,"analysis/01_Humans_GPT4_and_CA_performances/data/02_GPT4/GPT4_SLCA.rds")
